+++ application/npcs/views.py
@@ -27,7 +27,7 @@ def npcs_form(campaign_id):
 @app.route("/campaigns/<campaign_id>/npcs/<npc_id>/remove/", methods=["POST"])
 @login_required
 def remove_npc(npc_id, campaign_id):
-    if Campaign.is_registered_to_campaign(campaign_id, current_user):
+    if Campaign.is_registered_to_campaign(campaign_id, current_user) and Campaign.is_campaign_admin(campaign_id, current_user):
         db.session.query(Npc).filter_by(id=npc_id).delete()
         db.session().commit()
         return redirect(url_for("npcs_index", campaign_id=campaign_id))
@@ -64,7 +64,7 @@ def npcs_create(campaign_id):
 @login_required
 def open_npc(npc_id, campaign_id):
     if Campaign.is_registered_to_campaign(campaign_id, current_user):
-        return render_template("npcs/npc.html", npc=db.session.query(Npc).get(npc_id), campaign_id=campaign_id)
+        return render_template("npcs/npc.html", npc=db.session.query(Npc).get(npc_id), campaign_id=campaign_id, user_is_admin=Campaign.is_campaign_admin(campaign_id, current_user))
     return redirect(url_for("campaigns_register", campaign_id=campaign_id))
